JWT authentication added

+++ controllers/api/v1/post_api.js
@@ -1,5 +1,6 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
+const passport = require('passport');
 
 
 module.exports.index = async (req, res)=>{
@@ -25,18 +26,18 @@ module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
-
-        // if (post.user == req.user.id){
+        if (post && post.user == req.user._id.toString()){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
-
-
-    
             return res.status(200).json({
                 message: "Post and associated comments deleted successfully!"
             });
-      
+        }else{
+            return res.status(401).json({
+                message:"you cannot delete this post",
+            });
+        }
 
     }catch(err){
         console.log('********', err);
